+++ docs/changelog.md
@@ -8,6 +8,7 @@
 - Reduce memory use during congeneric series fitting in [#17](https://github.com/cole-group/presto/pull/17)
 - Improve test quality, speed, and coverage in [#19](https://github.com/cole-group/presto/pull/19)
 - Update environments with explicit python and cuda versions and document installation with different cuda versions
+- Plot sampling of rotatable torsions in [#23](https://github.com/cole-group/presto/pull/23)
 
 ### Fixes